Further refactoring of user-related modules
