+++ lib/sanbase/accounts/user/user_email.ex
@@ -7,6 +7,11 @@ defmodule Sanbase.Accounts.User.Email do
   require Mockery.Macro
 
   @token_valid_window_minutes 60
+  @email_token_length 64
+
+  def generate_email_token() do
+    :crypto.strong_rand_bytes(@email_token_length) |> Base.url_encode64()
+  end
 
   def find_by_email_candidate(email_candidate, email_candidate_token) do
     email_candidate = String.downcase(email_candidate)
@@ -26,7 +31,7 @@ defmodule Sanbase.Accounts.User.Email do
   def update_email_token(user, consent \\ nil) do
     user
     |> User.changeset(%{
-      email_token: User.generate_email_token(),
+      email_token: generate_email_token(),
       email_token_generated_at: NaiveDateTime.utc_now() |> NaiveDateTime.truncate(:second),
       email_token_validated_at: nil,
       consent_id: consent
@@ -38,7 +43,7 @@ defmodule Sanbase.Accounts.User.Email do
     user
     |> User.changeset(%{
       email_candidate: email_candidate,
-      email_candidate_token: User.generate_email_token(),
+      email_candidate_token: generate_email_token(),
       email_candidate_token_generated_at:
         NaiveDateTime.utc_now() |> NaiveDateTime.truncate(:second),
       email_candidate_token_validated_at: nil
